refactor : 277 PaymentService 제거, 직접 호출하도록 변경
