+++ resale/src/main/java/com/goti/resale/service/application/ResaleOrderEventListener.java
@@ -20,6 +20,9 @@ import com.goti.resale.domain.entity.resale.ResaleListingEntity;
 import com.goti.resale.domain.entity.resale.ResalePriceHistoryEntity;
 import com.goti.resale.domain.entity.resale.ResaleRestrictionEntity;
 import com.goti.resale.domain.entity.resale.ResaleTransactionEntity;
+import com.goti.resale.dto.request.ResalePaymentRequest;
+import com.goti.resale.dto.request.ResaleTransactionItemRequest;
+import com.goti.resale.infra.PaymentApiClient;
 import com.goti.resale.infra.TicketClient;
 import com.goti.resale.infra.dto.ResaleOrderCreatedEvent;
 import com.goti.resale.infra.dto.ResaleOrderPaymentCompletedEvent;
@@ -28,8 +31,7 @@ import com.goti.resale.repository.ResaleRestrictionRepository;
 import com.goti.resale.repository.ResaleTransactionRepository;
 import com.goti.resale.repository.history.ResalePriceHistoryRepository;
 import com.goti.resale.repository.listing.ResaleListingRepository;
-import com.goti.resale.service.domain.ResaleRestrictionService;
-import com.goti.resale.service.infra.PaymentService;
+import com.goti.resale.service.domain.ResaleRestrictionDomainService;
 import com.goti.resale.utils.ResaleRestrictionHandler;
 
 import lombok.RequiredArgsConstructor;
@@ -45,14 +47,14 @@ public class ResaleOrderEventListener {
 	private final ResalePriceHistoryRepository priceHistoryRepository;
 	private final ResaleRestrictionRepository restrictionRepository;
 	private final ResaleRestrictionHandler restrictionHandler;
-	private final ResaleRestrictionService restrictionService;
-	private final PaymentService paymentService;
+	private final ResaleRestrictionDomainService restrictionDomainService;
+	private final PaymentApiClient paymentApiClient;
 	private final TicketClient ticketClient;
 
 	@Async
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
 	public void handleOrderCreated(ResaleOrderCreatedEvent event) {
-		paymentService.createResalePayment(
+		createResalePayment(
 			event.orderId(),
 			event.buyerId(),
 			event.totalBuyerAmount(),
@@ -73,7 +75,7 @@ public class ResaleOrderEventListener {
 		List<ResaleListingEntity> resaleListings = new ArrayList<>();
 		List<ResalePriceHistoryEntity> priceHistories = new ArrayList<>();
 
-		ResaleRestrictionEntity restriction = restrictionService.getOrCreateRestriction(event.buyerId());
+		ResaleRestrictionEntity restriction = restrictionDomainService.getOrCreateRestriction(event.buyerId());
 
 		for (ResaleTransactionEntity transaction : transactions) {
 			ResaleListingEntity resaleListing = transaction.getListing();
@@ -109,7 +111,7 @@ public class ResaleOrderEventListener {
 			transferOwnershipAsync(listing.getTicketId(), event.buyerId());
 		}
 
-		paymentService.releaseEscrow(event.resaleOrderId());
+		releaseEscrow(event.resaleOrderId());
 	}
 
 	@TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
@@ -140,6 +142,44 @@ public class ResaleOrderEventListener {
 			log.error("티켓 소유권 이전 실패 - 티켓ID: {}, 구매자: {}, 에러: {}",
 				ticketId, buyerId, e.getMessage(), e);
 			throw new CustomException(ErrorCode.TRANSFER_OWNERSHIP_FAILED);
+		}
+	}
+
+	public void createResalePayment(
+		UUID orderId,
+		UUID buyerId,
+		int totalAmount,
+		int totalBuyerFee,
+		int totalSellerFee,
+		List<ResaleTransactionItemRequest> items,
+		String idempotencyKey
+	) {
+		ResalePaymentRequest request = new ResalePaymentRequest(
+			orderId,
+			buyerId,
+			totalAmount,
+			totalBuyerFee,
+			totalSellerFee,
+			items,
+			"CARD",
+			idempotencyKey
+		);
+
+		try {
+			paymentApiClient.createResalePayment(request);
+		} catch (Exception e) {
+			log.error("리셀 결제 요청 실패 - orderId: {}, error: {}", orderId, e.getMessage());
+			throw new CustomException(ErrorCode.RESALE_PAYMENT_FAILED);
+		}
+	}
+
+	public void releaseEscrow(UUID orderId) {
+		log.info("에스크로 해제 요청 - orderId: {}", orderId);
+		try {
+			paymentApiClient.releaseEscrow(orderId);
+		} catch (Exception e) {
+			log.error("에스크로 해제 실패 - orderId: {}, error: {}", orderId, e.getMessage());
+			throw new CustomException(ErrorCode.RESALE_ESCROW_FAILED);
 		}
 	}
 }
